improvement: renamed reqUtils to iputils
